+++ src/i18n.js
@@ -1,25 +1,18 @@
 /*eslint no-cond-assign: 0*/
-import i18n from 'i18next';
-import {assignObjectToKeys} from './utils';
+import * as i18n from 'i18next';
 
 export class I18N {
 
   globalVars = {};
 
-  constructor(ea) {
+  constructor(ea, signaler) {
     this.i18next = i18n;
     this.ea = ea;
     this.Intl = window.Intl;
-
-    // check whether Intl is available, otherwise load the polyfill
-    if (window.Intl === undefined) {
-      System.import('Intl').then( (poly) => {
-        window.Intl = poly;
-      });
-    }
+    this.signaler = signaler;
   }
 
-  setup(options) {
+  setup(options?) {
     const defaultOptions = {
       resGetPath: 'locale/__lng__/__ns__.json',
       lng: 'en',
@@ -43,6 +36,7 @@ export class I18N {
       let oldLocale = this.getLocale();
       this.i18next.setLng(locale, tr => {
         this.ea.publish('i18n:locale:changed', { oldValue: oldLocale, newValue: locale });
+        this.signaler.signal('aurelia-translation-signal');
         resolve(tr);
       });
     });
@@ -52,22 +46,40 @@ export class I18N {
     return this.i18next.lng();
   }
 
-  nf(options, locales) {
+  nf(options?, locales?) {
     return new this.Intl.NumberFormat(locales || this.getLocale(), options || {});
   }
 
-  df(options, locales) {
+  uf(number, locale?) {
+    let nf = this.nf({}, locale || this.getLocale());
+    let comparer = nf.format(10000 / 3);
+
+    let thousandSeparator = comparer[1];
+    let decimalSeparator  = comparer[5];
+
+    // remove thousand seperator
+    let result = number.replace(thousandSeparator, '')
+      // remove non-numeric signs except -> , .
+      .replace(/[^\d.,-]/g, '')
+      // replace original decimalSeparator with english one
+      .replace(decimalSeparator, '.');
+
+    // return real number
+    return Number(result);
+  }
+
+  df(options?, locales?) {
     return new this.Intl.DateTimeFormat(locales || this.getLocale(), options);
   }
 
-  tr(key, options) {
+  tr(key, options?) {
     let fullOptions = this.globalVars;
 
     if (options !== undefined) {
       fullOptions = Object.assign(Object.assign({}, this.globalVars), options);
     }
 
-    return this.i18next.t(key, assignObjectToKeys('', fullOptions));
+    return this.i18next.t(key, fullOptions);
   }
 
   registerGlobalVariable(key, value) {