+++ src/t.js
@@ -1,6 +1,8 @@
 import {I18N} from './i18n';
 import {EventAggregator} from 'aurelia-event-aggregator';
 import {customAttribute} from 'aurelia-templating';
+import {SignalBindingBehavior} from 'aurelia-templating-resources';
+import {ValueConverter} from 'aurelia-binding';
 import {LazyOptional} from './utils';
 
 
@@ -43,8 +45,9 @@ export class TCustomAttribute {
 
   bind() {
     this.params = this.lazyParams();
+    this.timers = [];
 
-    setTimeout( () => {
+    this.timers.push(setTimeout( () => {
       if (this.params) {
         this.params.valueChanged = (newParams, oldParams) => {
           this.paramsChanged(this.value, newParams, oldParams);
@@ -56,10 +59,10 @@ export class TCustomAttribute {
         this.service.updateValue(this.element, this.value, p);
       });
 
-      setTimeout( () => {
+      this.timers.push(setTimeout( () => {
         this.service.updateValue(this.element, this.value, p);
-      });
-    });
+      }));
+    }));
   }
 
   paramsChanged(newValue, newParams) {
@@ -72,6 +75,42 @@ export class TCustomAttribute {
   }
 
   unbind() {
-    this.subscription.dispose();
+    // Clear timers so that we do not run unecessary code after unbinding
+    this.timers.forEach(t => clearTimeout(t));
+    // If unbind is called before timeout for subscription is triggered, subscription will be undefined
+    if (this.subscription) {
+      this.subscription.dispose();
+    }
+  }
+}
+
+export class TBindingBehavior {
+  static inject = [SignalBindingBehavior];
+
+  constructor(signalBindingBehavior) {
+    this.signalBindingBehavior = signalBindingBehavior;
+  }
+
+  bind(binding, source) {
+    // bind the signal behavior
+    this.signalBindingBehavior.bind(binding, source, 'aurelia-translation-signal');
+
+    // rewrite the expression to use the TValueConverter.
+    // pass through any args to the binding behavior to the TValueConverter
+    let sourceExpression = binding.sourceExpression;
+    let expression = sourceExpression.expression;
+    sourceExpression.expression = new ValueConverter(
+      expression,
+      't',
+      sourceExpression.args,
+      [expression, ...sourceExpression.args]);
+  }
+
+  unbind(binding, source) {
+    // undo the expression rewrite
+    binding.sourceExpression.expression = binding.sourceExpression.expression.expression;
+
+    // unbind the signal behavior
+    this.signalBindingBehavior.unbind(binding, source);
   }
 }