+++ test/unit/i18n-attribute.spec.js
@@ -1,25 +1,39 @@
-import {initialize, DOM} from 'aurelia-pal-browser';
+import {DOM} from 'aurelia-pal';
+import {BindingSignaler} from 'aurelia-templating-resources';
+import {initialize} from 'aurelia-pal-browser';
 import {TCustomAttribute, TParamsCustomAttribute} from '../../src/t';
 import {Container} from 'aurelia-dependency-injection';
-import {templatingEngine} from 'aurelia-templating';
+import {TemplatingEngine} from 'aurelia-templating';
 import {I18N} from '../../src/i18n';
+import {EventAggregator} from 'aurelia-event-aggregator';
 
 initialize();
 
 describe('testing i18n attributes', () => {
   let container;
+  let templatingEngine;
+  let sut;
 
   beforeEach(() => {
-    container = new Container();
-    let element = DOM.createElement('div');
-    container.registerInstance(DOM.Element, element);
+    sut = new I18N(new EventAggregator(), new BindingSignaler());
+    sut.setup({
+      resStore: {},
+      lng: 'en',
+      getAsync: false,
+      sendMissing: false,
+      fallbackLng: 'en',
+      debug: false
+    });
 
-    templatingEngine.initialize(container);
+    container = new Container();
+    container.registerInstance(I18N, sut);
+    container.registerInstance(DOM.Element, DOM.createElement('div'));
+    templatingEngine = container.get(TemplatingEngine);
   });
 
 
   it('should raise value change on i18n custom attribute', done => {
-    let i18nAttribute = templatingEngine.createModelForUnitTest(TCustomAttribute);
+    let i18nAttribute = templatingEngine.createViewModelForUnitTest(TCustomAttribute);
     spyOn(i18nAttribute, 'valueChanged');
 
     // disable DOM operations by mocking the specific function
@@ -34,10 +48,10 @@ describe('testing i18n attributes', () => {
   });
 
   it('should listen to params changes', done => {
-    let paramsAttribute = templatingEngine.createModelForUnitTest(TParamsCustomAttribute);
+    let paramsAttribute = templatingEngine.createViewModelForUnitTest(TParamsCustomAttribute);
     container.registerInstance(TParamsCustomAttribute, paramsAttribute);
 
-    let i18nAttribute = templatingEngine.createModelForUnitTest(TCustomAttribute);
+    let i18nAttribute = templatingEngine.createViewModelForUnitTest(TCustomAttribute);
     spyOn(i18nAttribute, 'paramsChanged').and.callThrough();
     i18nAttribute.bind();